Split test build graph command tests
